+++ src/engines/hfbi/full/mod.rs
@@ -34,7 +34,7 @@ pub fn calculate_mmi(
     let bn: f32 = calc_bn(campionamento);
     let dbent: f32 = calc_dbent(campionamento, anagrafica)?;
     let ddom: f32 = calc_ddom(campionamento, anagrafica)?;
-    let dhzp: f32 = calc_dhzp(campionamento, anagrafica);
+    let dhzp: f32 = calc_dhzp(campionamento, anagrafica)?;
     let dmig: f32 = calc_dmig(campionamento, anagrafica);
     let rqe_bbent = bbent / condizioni_riferimento.bbent;
     let rqe_bn = bn / condizioni_riferimento.bn;